backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.0
youtube-transcript-api==1.2.4
google-generativeai==0.8.3
python-dotenv==1.0.1
